requirement.txt
Django==1.8.6
django-hvad==1.3.0
psycopg2==2.6.1
djangorestframework==3.2.4
django-rosetta==0.7.6